pyproject.toml
[tool.poetry]
name = "thesis-sketch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
tqdm = "^4.64.1"
matplotlib = "^3.6.3"
seaborn = "^0.12.2"
pandas = "^1.5.3"
numpy = "^1.24.1"
rank-bm25 = "^0.2.2"
scikit-learn = "^1.2.1"
xmltodict = "^0.13.0"
pytorch-lightning = "^1.9.0"
joblib = "^1.2.0"
scrapy = "^2.7.1"
torch = "^1.13.1"
transformers = "^4.26.0"
hydra-core = "^1.3.1"
pyrootutils = "^1.0.4"
rich = "^13.3.1"
torchmetrics = "^0.11.1"
wandb = "^0.13.9"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
ruff = "^0.0.237"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 120
target-version = ['py39', 'py310']
include = '\.pyi?$'


[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.mccabe]
max-complexity = 10
